third_party/blink/web_tests/svg/hixie/transform/001-expected.txt

CONSOLE ERROR: Error: <circle> attribute transform: Unexpected end of attribute. Expected ')', "translate(200,0".
